+++ doomgeneric/p_map.c
@@ -535,7 +535,7 @@ P_TryMove
 	    if (side != oldside)
 	    {
 		if (ld->special)
-		    P_CrossSpecialLine (ld-lines, oldside, thing);
+		    P_CrossSpecialLine ((int) (ld-lines), oldside, thing);
 	    }
 	}
     }
@@ -1422,7 +1422,7 @@ static void SpechitOverrun(line_t *ld)
     
     // Calculate address used in doom2.exe
 
-    addr = baseaddr + (ld - lines) * 0x3E;
+    addr = baseaddr + (unsigned int) ((ld - lines) * 0x3E);
 
     switch(numspechit)
     {
