+++ doomgeneric/p_maputl.c
@@ -600,7 +600,7 @@ PIT_AddLineIntercepts (line_t* ld)
     intercept_p->frac = frac;
     intercept_p->isaline = true;
     intercept_p->d.line = ld;
-    InterceptsOverrun(intercept_p - intercepts, intercept_p);
+    InterceptsOverrun((int) (intercept_p - intercepts), intercept_p);
     intercept_p++;
 
     return true;	// continue
@@ -666,7 +666,7 @@ boolean PIT_AddThingIntercepts (mobj_t* thing)
     intercept_p->frac = frac;
     intercept_p->isaline = false;
     intercept_p->d.thing = thing;
-    InterceptsOverrun(intercept_p - intercepts, intercept_p);
+    InterceptsOverrun((int) (intercept_p - intercepts), intercept_p);
     intercept_p++;
 
     return true;		// keep going
@@ -688,7 +688,7 @@ P_TraverseIntercepts
     intercept_t*	scan;
     intercept_t*	in;
 	
-    count = intercept_p - intercepts;
+    count = (int) (intercept_p - intercepts);
     
     in = 0;			// shut up compiler warning
 	
@@ -846,7 +846,7 @@ static void InterceptsOverrun(int num_intercepts, intercept_t *intercept)
 
     InterceptsMemoryOverrun(location, intercept->frac);
     InterceptsMemoryOverrun(location + 4, intercept->isaline);
-    InterceptsMemoryOverrun(location + 8, (int) intercept->d.thing);
+    InterceptsMemoryOverrun(location + 8, (int) ((intptr_t) intercept->d.thing));
 }
